Some more fixes

+++ tools/create_tag.sh
@@ -10,14 +10,18 @@ createTag() {
         echo "ERROR: no tag given"
         exit 1
     fi
-    git clone -b ${BRANCH_TO_TAG} $(git remote get-url origin) tagging
-    cd tagging
+    if "$3" == "clean-checkout" ]] ; then
+        git clone -b ${BRANCH_TO_TAG} $(git remote get-url origin) tagging
+        cd tagging
+    fi
     NEW_TAG="$2"
     echo "Tagging branch ${BRANCH_TO_TAG} with ${NEW_TAG}"
     git tag -a -m "Created tag ${NEW_TAG}" "${NEW_TAG}"
     git push -u origin "${NEW_TAG}"
-    cd ..
-    rm -rf tagging
+    if "$3" == "clean-checkout" ]] ; then
+        cd ..
+        rm -rf tagging
+    fi
 }
 
 getReleaseType() {
@@ -69,15 +73,15 @@ elif [[ "${BRANCH_NAME}" == "master" ]]; then
         NEW_DEVELOP_VERSION="${VNUM1}.${VNUM2}.0-dev.0"
         echo "Creating new minor release ${NEW_MASTER_VERSION}"
         createTag "master" "${NEW_MASTER_VERSION}"
-        createTag "develop" "${NEW_DEVELOP_VERSION}"
+        createTag "develop" "${NEW_DEVELOP_VERSION}" "clean-checkout"
     elif [[ "${RELEASE_TYPE}" == "release-major" ]]; then
-        VNUM1="$((VNUM1+1))"
+        VNUM1="v$((VNUM1+1))"
         #create new tag
         NEW_MASTER_VERSION="${VNUM1}.0.1"
         NEW_DEVELOP_VERSION="${VNUM1}.0.0-dev.0"
         echo "Creating new major release ${NEW_MASTER_VERSION}"
         createTag "master" "${NEW_MASTER_VERSION}"
-        createTag "develop" "${NEW_DEVELOP_VERSION}"
+        createTag "develop" "${NEW_DEVELOP_VERSION}" "clean-checkout"
     else
         echo "INFO: this commit has no release-patch, release-minor or release-major specified, not creating a new release version"
         exit 0
@@ -86,4 +90,3 @@ else
     echo "ERROR: this script is only allowed to be called for branches develop and master"
     exit 1
 fi
-git pull
